Logic class

+++ src/Event.java
@@ -11,7 +11,7 @@ public class Event {
 	private Vector<Hotel> suggestedHotels;
 	private Vector<Programme> programmeSchedule;
 	private EventType eventType;
-	private float eventBudget;
+	private double eventBudget;
 	private String eventDescription;
 	
 	public Event(){
@@ -96,11 +96,11 @@ public class Event {
 		this.programmeSchedule = programmeSchedule;
 	}
 
-	public float getEventBudget() {
+	public double getEventBudget() {
 		return eventBudget;
 	}
 
-	public void setEventBudget(float eventBudget) {
+	public void setEventBudget(double eventBudget) {
 		this.eventBudget = eventBudget;
 	}
